Centralise domain lookup
Previously, lookup of hosts was performed in both the http worker, for
new connections, and the lookup worker, for crawled urls.

Centralise all domain lookup of hosts to the lookup worker.
Remove multiple hostnames per connection, as due to TLS in HTTPS a new
connection will be created for a different hostname anyway.