Support HTTP 429 request rate-limiting
Not all servers specify Crawl-delay in robots.txt, so implement
a simple implementation of HTTP 429 rate-limiting to avoid too
many requests being sent to those servers.

This is a simple implementation as it never re-requests the
rate-limited resource and does not consider that HTTP 429
rate-limiting can be done on a per-resource basis, but treats
any HTTP 429 as covering the entire host:port.
This is enough to avoid too many requests being sent to a server.